Renamed SUn.h -> GaugeGroup.h
